% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_content.R
\name{package_f2}
\alias{package_f2}
\title{Two roads diverged in a yellow wood,
And sorry I could not travel both
And be one traveler, long I stood
And looked down one as far as I could
To where it bent in the undergrowth;}
\usage{
package_f2(a, b)
}
\arguments{
\item{a}{The best input ever}

\item{b}{The awesome second input}
}
\description{
Then took the other, as just as fair,
And having perhaps the better claim,
Because it was grassy and wanted wear;
Though as for that the passing there
Had worn them really about the same,
}
\details{
And both that morning equally lay
In leaves no step had trodden black.
Oh, I kept the first for another day!
Yet knowing how way leads on to way,
I doubted if I should ever come back.

I shall be telling this with a sigh
Somewhere ages and ages hence:
Two roads diverged in a wood, and I—
I took the one less traveled by,
And that has made all the difference.
}
